LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-dynamic.tentative-expected.txt


Harness Error (TIMEOUT), message = null

FAIL Switching between document and scroll timelines [immediate] assert_equals: expected "120px" but got "0px"
FAIL Switching between document and scroll timelines [scroll] assert_equals: expected "120px" but got "0px"
TIMEOUT Switching pending animation from document to scroll timelines [immediate] Test timed out
NOTRUN Switching pending animation from document to scroll timelines [scroll]
NOTRUN Changing computed value of animation-timeline changes effective timeline [immediate]
NOTRUN Changing computed value of animation-timeline changes effective timeline [scroll]
NOTRUN Changing to/from animation-timeline:none [immediate]
NOTRUN Changing to/from animation-timeline:none [scroll]
NOTRUN Reverse animation direction [immediate]
NOTRUN Reverse animation direction [scroll]
NOTRUN Change to timeline attachment while paused [immediate]
NOTRUN Change to timeline attachment while paused [scroll]
NOTRUN Switching timelines and pausing at the same time [immediate]
NOTRUN Switching timelines and pausing at the same time [scroll]
